主页增加关闭页面按钮 去掉html包,增加window平台 处理一些在window上运行报错问题

+++ lib/app/utils/ui/auto_ui.dart
@@ -15,12 +15,12 @@ num setHeight(num size){
 //适配文字
 num setSp(num size){
 //  return ScreenUtil().setSp(size);
-  return size;
+  return size.toDouble();
 }
 
 //自动适配,后面方便扩展
 num auto(num size){
-  return size;
+  return size.toDouble();
 }
